Move BinOp and Builtin definitions in the relevant module
